requirements.txt
mysql-connector-python~=8.0.27
translate~=3.6.1
python-telegram-bot=13.3
Telethon~=1.24.0